+++ CHANGES.rst
@@ -3,6 +3,8 @@ Version 0.15.0
 
 Unreleased
 
+- Add ``ThreadedSimpleCache``, a thread-safe variant of ``SimpleCache`` backed
+  by a reentrant ``threading.RLock``. :issue:`446`
 
 Version 0.14.0
 --------------